+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.17.0 (2022-08-18)
+
+### Feat
+
+- **models**: introduce smart serializer functions to serialize both model and queryset
+
+## v0.16.1 (2022-08-17)
+
+### Fix
+
+- docs return no api  response schema
+
 ## v0.16.0 (2022-08-16)
 
 ### Feat